Use gcc function attribute `format' to catch bad format strings.  Fix them.  From Marc Olzheim. Type modifier 'l' was missing in many places, probably rendering the server useless on 64-bit hosts. (ef_flush, ef_write, ef_extend, lwpCreate, lwpDestroy): Use conversion specifier 'p' to print pointers. (check_market): Fix display of loan amount. (doland): Fix unescaped '%' (undefined behavior). (ldump, ndump, pdump, sdump): Don't use flag '0' with conversion specifier 's' (undefined behavior). (dump, ldump, lost, ndump, pdump, sdump, empth_create, update_sched): Cast time_t and pthread_t to long for printing. (lwpStackCheck, lwpStackCheckUsed, finish_sects): Insert cast to fix argument type on all platforms. (prod): Remove extra argument. (perform_mission, airdamage, retreat_land1, lwpReschedule): Format string missed arguments.

+++ src/lib/commands/dump.c
@@ -248,7 +248,7 @@ dump(void)
     if (player->god)
 	pr("   ");
     time(&now);
-    pr("DUMP SECTOR %d\n", now);
+    pr("DUMP SECTOR %ld\n", (long)now);
     if (player->god)
 	pr("own ");
     pr("x y");
